split traversal from reporting
